refactor: extract data_ops and utility unit tests into submodules
Per #26: first incremental step — move #[cfg(test)] modules to
data_ops/tests.rs and utility/tests.rs so production code compiles
separately and files shrink for review. Further API-cluster splits
can follow without carrying multi-kLOC test modules.

Fixes #26
